# prov/rxm: Optimizations of the critical path