This patch adds some optimizations for the  RxM critical path
- Inject and tagged inject paths are splitted into special function.
- Optimizations of CQ engine

**NOTE:**
The part of this PR contains some commits form #3686 that will be removed after #3686 is being merged into the master branch:
- Avoid excessive variable
- Minor changes to the RxM and utility/cmap code. This reduces only single instruction on critical path when the library is built by GCC
- RXM_NONE state has been removed as unneeded and used only in debug case
- Make TX/RX entry routines as static inline
- Simplifies conditions for MR_LOCAL and non-MR_LOCAL cases
- rxm_pkt_init() is doing an excessive memset()
- Remove function's declarations that are defined, but aren't implemented
- Mark some local functions as static