@gladkovdmitry17 2018-01-09
@a-ilango Could you take a look, please? Do you find this patch viable for the RxM provider?