@gladkovdmitry17 2018-01-10
These are instruction flows for the MPICH-CH3 over
- [OFI/RxM master branch](https://github.com/ofiwg/libfabric/files/1619111/ofi_rxm_master.txt)
- [OFI/RxM gladkovmitry17:rxm branch](https://github.com/ofiwg/libfabric/files/1619112/ofi_rxm_master_opt.txt)