@gladkovdmitry17 2018-01-11
@shefty @a-ilango I've updated PR splitting commits to the small ones. Seems it should be more easy to review.
Also I'm attaching to this comment the results of IMB (latency) and instructions counting.
- IMB
```
       #bytes new  old
            0 1.55 1.60
            1 1.56 1.61
            2 1.56 1.61
            4 1.56 1.61
            8 1.56 1.61
           16 1.56 1.61
           32 1.56 1.61
           64 1.61 1.66
          128 1.65 1.69
          256 2.14 2.16
          512 2.28 2.31
         1024 2.46 2.50
         2048 2.94 2.97
         4096 3.88 4.00
         8192 5.45 5.70
        16384 8.20 8.21
```
- instruction counting of send path of
 [OLD version](https://github.com/ofiwg/libfabric/files/1622656/ofi_rxm_old.txt) (MPICH + OFI = 1117)
 [NEW version](https://github.com/ofiwg/libfabric/files/1622657/ofi_rxm_new.txt) (MPICH + OFI = 888)