@gladkovdmitry17 2018-01-12
I've pushed some new comments.
The last commit reduces the amount of service data. The previous size of service data is 64 bytes (= cache line size). Now the size of service data is 56 bytes. This optimizes sending of buffers that has size <= 8 bytes (it touches only one cache line instead). Btw, A buffer of 8 bytes is the most popular size for applications on top of MPI

These are gained new latency results:
```
       #bytes new  old
            0 1.50 1.60
            1 1.51 1.61
            2 1.51 1.61
            4 1.51 1.61
            8 1.51 1.61
           16 1.52 1.61
           32 1.51 1.61
           64 1.55 1.66
          128 1.59 1.69
          256 2.04 2.16
          512 2.16 2.31
         1024 2.37 2.50
         2048 2.80 2.97
         4096 3.72 4.00
         8192 5.40 5.70
        16384 8.11 8.21
```

These are the # of instructions for:
- RxM from this branch: [Send](https://github.com/ofiwg/libfabric/files/1626568/ofi_rxm_send_new.txt)(864) and [Recv](https://github.com/ofiwg/libfabric/files/1626567/ofi_rxm_recv_new.txt)(1992) paths
- RxM from the master branch: [Send](https://github.com/ofiwg/libfabric/files/1626569/ofi_rxm_send_old.txt)(1117) and [Recv](https://github.com/ofiwg/libfabric/files/1626570/ofi_rxm_recv_old.txt)(2218) paths