@gladkovdmitry17 2018-01-18
@shefty @a-ilango I've refreshed the content of this PR by applying your comments and adding two more commits. Could look into this PR, please? Thanks in advance!