@gladkovdmitry17 2018-01-18
> I gave this only a quick review.

@shefty I;ve covered your comments by appropriate fixes