@gladkovdmitry17 2018-01-18
@a-ilango @shefty Thanks for taking a look into this!
I've prepared a new version of the PR w/ applied remarks