@gladkovdmitry17 2018-01-19
@jswaro @hppritcha Something is wrong with LANL-CI for all PRs.

Ignoring Travis CI OSX jobs hanging and LANL-CI failure as is not connected with this PR